Create a separate TimeTable component
