+++ src/panel/poi/PoiBlockContainer.jsx
@@ -28,7 +28,7 @@ export default class PoiBlockContainer extends React.Component {
 
     return <div className="poi_panel__info">
       {displayCovidInfo && <CovidBlock block={covidBlock} normalHourBlock={hourBlock} />}
-      {hourBlock && <HourBlock block={hourBlock} covid19enabled={displayCovidInfo} />}
+      {hourBlock && <HourBlock block={hourBlock} covid19enabled={!!displayCovidInfo} />}
       {informationBlock && <InformationBlock block={informationBlock} />}
       {websiteBlock && <WebsiteBlock block={websiteBlock} poi={this.props.poi} />}
       {contactBlock && <ContactBlock block={contactBlock} />}
