+++ src/panel/poi/blocks/Hour.jsx
@@ -1,30 +1,8 @@
 /* global _ */
 import React from 'react';
 import OsmSchedule from 'src/adapters/osm_schedule';
-import classnames from 'classnames';
+import TimeTable from './TimeTable';
 import PropTypes from 'prop-types';
-
-function showHour(day) {
-  if (day.opening && day.opening.length > 0) {
-    return day.opening.map((openingFragment, i) =>
-      <p key={i}>{ openingFragment.beginning } - { openingFragment.end }</p>);
-  }
-  return _('Closed', 'hour block');
-}
-
-function showHours(displayHours) {
-  const dayNumber = new Date().getDay();
-
-  return <tbody>
-    {displayHours.map((day, i) =>
-      <tr key={i} className={
-        classnames({ 'poi_panel__info__hours--current': (i + 1) % 7 === dayNumber })
-      }>
-        <td className="day">{ day.dayName }</td>
-        <td className="hours">{ showHour(day) }</td>
-      </tr>)}
-  </tbody>;
-}
 
 function renderTitle(opening, covid19) {
   if (covid19) {
@@ -52,7 +30,6 @@ export default class HourBlock extends React.Component {
 
   constructor(props) {
     super(props);
-    this.state = { isCollapsed: true };
 
     this.messages = {
       open: {
@@ -64,42 +41,6 @@ export default class HourBlock extends React.Component {
         color: '#8c0212',
       },
     };
-
-    this.expandCollapse = () => {
-      this.setState(state => ({
-        isCollapsed: !state.isCollapsed,
-      }));
-    };
-  }
-
-  renderStatus(opening) {
-    // TODO: use OpeningHour instead (careful! OsmSchedule initialization happens there as well!)
-    if (opening.isTwentyFourSeven) {
-      return <div className="poi_panel__info__hours__status__text poi_panel__info__hours__24_7">
-        { _('Open 24/7', 'hour block') }
-        <div className="poi_panel__info__hour__circle"
-          style={{ background: opening.status.color }}
-        />
-      </div>;
-    }
-    return <div>
-      <div className="poi_panel__info__hours__status" onClick={this.expandCollapse}>
-        { renderTitle(opening, this.props.covid19enabled) }
-        <i className={classnames(
-          'icon-icon_chevron-down',
-          'poi_panel__info__hours__status__toggle',
-          {
-            'poi_panel__info__hours__status__toggle--reversed': !this.state.isCollapsed,
-          })} />
-      </div>
-      <div className={classnames('poi_panel__info__hours', {
-        'poi_panel__info__hours--open': !this.state.isCollapsed,
-      })}>
-        <table className="poi_panel__info__hours__table">
-          { showHours(opening.displayHours) }
-        </table>
-      </div>
-    </div>;
   }
 
   render() {
@@ -112,7 +53,7 @@ export default class HourBlock extends React.Component {
       <div className="poi_panel__info__section__description">
         <div className="icon-icon_clock poi_panel__block__symbol"></div>
         <div className="poi_panel__block__content">
-          { this.renderStatus(opening) }
+          <TimeTable title={renderTitle(opening, this.props.covid19enabled)} schedule={opening} />
         </div>
       </div>
     </div>;
